requirements.txt
alembic==1.4.3
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
click==7.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Login==0.5.0
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
importlib-metadata==2.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==4.6.11
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==1.2.0
